+++ hazelcast/src/main/java/com/hazelcast/spi/impl/BasicInvocationBuilder.java
@@ -26,19 +26,7 @@ import com.hazelcast.spi.Operation;
 /**
  * An {@link com.hazelcast.spi.InvocationBuilder} that is tied to the {@link com.hazelcast.spi.impl.BasicOperationService}.
  */
-public class BasicInvocationBuilder implements InvocationBuilder {
-
-    private final NodeEngineImpl nodeEngine;
-    private final String serviceName;
-    private final Operation op;
-    private final int partitionId;
-    private final Address target;
-    private Callback<Object> callback;
-    private long callTimeout = -1L;
-    private int replicaIndex = 0;
-    private int tryCount = 250;
-    private long tryPauseMillis = 500;
-    private String executorName;
+public class BasicInvocationBuilder extends InvocationBuilder {
 
     public BasicInvocationBuilder(NodeEngineImpl nodeEngine, String serviceName, Operation op, int partitionId) {
         this(nodeEngine, serviceName, op, partitionId, null);
@@ -50,101 +38,7 @@ public class BasicInvocationBuilder implements InvocationBuilder {
 
     private BasicInvocationBuilder(NodeEngineImpl nodeEngine, String serviceName, Operation op,
                                    int partitionId, Address target) {
-        this.nodeEngine = nodeEngine;
-        this.serviceName = serviceName;
-        this.op = op;
-        this.partitionId = partitionId;
-        this.target = target;
-    }
-
-    @Override
-    public String getExecutorName() {
-        return executorName;
-    }
-
-    @Override
-    public InvocationBuilder setExecutorName(String executorName) {
-        this.executorName = executorName;
-        return this;
-    }
-
-    @Override
-    public InvocationBuilder setReplicaIndex(int replicaIndex) {
-        if (replicaIndex < 0 || replicaIndex >= PartitionView.MAX_REPLICA_COUNT) {
-            throw new IllegalArgumentException("Replica index is out of range [0-"
-                    + (PartitionView.MAX_REPLICA_COUNT - 1) + "]");
-        }
-        this.replicaIndex = replicaIndex;
-        return this;
-    }
-
-    @Override
-    public InvocationBuilder setTryCount(int tryCount) {
-        this.tryCount = tryCount;
-        return this;
-    }
-
-    @Override
-    public InvocationBuilder setTryPauseMillis(long tryPauseMillis) {
-        this.tryPauseMillis = tryPauseMillis;
-        return this;
-    }
-
-    @Override
-    public InvocationBuilder setCallTimeout(long callTimeout) {
-        this.callTimeout = callTimeout;
-        return this;
-    }
-
-    @Override
-    public String getServiceName() {
-        return serviceName;
-    }
-
-    @Override
-    public Operation getOp() {
-        return op;
-    }
-
-    @Override
-    public int getReplicaIndex() {
-        return replicaIndex;
-    }
-
-    @Override
-    public int getTryCount() {
-        return tryCount;
-    }
-
-    @Override
-    public long getTryPauseMillis() {
-        return tryPauseMillis;
-    }
-
-    @Override
-    public Address getTarget() {
-        return target;
-    }
-
-    @Override
-    public int getPartitionId() {
-        return partitionId;
-    }
-
-    @Override
-    public long getCallTimeout() {
-        return callTimeout;
-    }
-
-    @Override
-    public Callback getCallback() {
-        return callback;
-    }
-
-    @Override
-    public InvocationBuilder setCallback(Callback<Object> callback) {
-        this.callback = callback;
-        return this;
+        super(nodeEngine,serviceName,op,partitionId,target);
     }
 
     @Override
